Separate tests.
