+++ tests/test_assetra.py
@@ -1,0 +1,10 @@
+import unittest 
+import sys
+from pathlib import Path
+
+sys.path.append('..')
+
+from powersys import energy_source
+
+if __name__ == '__main__':
+	unittest.main()
